meta: move the file writer into a separate function
Previously, cio_meta_write() contained a lot of filesystem specific
codes and was hard to support CIO_HAVE_BACKEND_FILESYSTEM=Off.

This patch separates out filesystem-dependant stuffs cleanly, and at the
same time enables us to compile cio_meta.c on Windows (VC++) nicely.

NOTE: This patch just moves a chunk of codes, involiving no actual
change to the logic.
